requirements.txt
requests
github-heatmap
python-dotenv